+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.0 under development
 
 - Enh #617: Add debug collector for yiisoft/yii-debug (@xepozz)
+- Chg #722: Remove legacy array syntax for typecast. Use `Param` instead (@terabytesoftw)
 
 ## 1.0.0 April 12, 2023